+++ extensions/api/identity-api/keypair-api/src/main/java/org/eclipse/edc/identityhub/api/keypair/KeyPairResourceManagementApiExtension.java
@@ -16,7 +16,7 @@ package org.eclipse.edc.identityhub.api.keypair;
 
 import org.eclipse.edc.identityhub.api.keypair.v1.unstable.GetAllKeyPairsApiController;
 import org.eclipse.edc.identityhub.api.keypair.v1.unstable.KeyPairResourceApiController;
-import org.eclipse.edc.identityhub.api.v1.validation.KeyDescriptorValidator;
+import org.eclipse.edc.identityhub.api.keypair.validation.KeyDescriptorValidator;
 import org.eclipse.edc.identityhub.spi.AuthorizationService;
 import org.eclipse.edc.identityhub.spi.IdentityHubApiContext;
 import org.eclipse.edc.identityhub.spi.keypair.KeyPairService;
