+++ extensions/api/identity-api/participant-context-api/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
     api(project(":spi:identity-hub-spi"))
     api(project(":spi:did-spi"))
     implementation(project(":extensions:api:identity-api:api-configuration"))
-    implementation(project(":extensions:api:identity-api:validators"))
+    implementation(project(":extensions:api:identity-api:validators:participant-context-validators"))
     implementation(libs.edc.spi.validator)
     implementation(libs.edc.spi.web)
     implementation(libs.edc.lib.util)
