+++ extensions/api/identity-api/participant-context-api/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.edc.identityhub.api.participantcontext.ParticipantContextManagementApiExtension
+org.eclipse.edc.identityhub.api.verifiablecredential.ParticipantContextManagementApiExtension
